+++ src/main/java/com/kryptokrauts/codegen/datatypes/ListMapper.java
@@ -1,39 +1,89 @@
 package com.kryptokrauts.codegen.datatypes;
 
+import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
+import com.squareup.javapoet.ParameterizedTypeName;
+import com.squareup.javapoet.TypeName;
+import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import java.lang.reflect.Type;
-import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
-/**
- * @TODO support Lists, Maps, Tuples, Nesting using a recursive approach
- *
- * @author mitch
- */
-public class ListMapper extends AbstractSophiaTypeMapper {
+public class ListMapper extends AbstractDatatypeMapper {
+
+  private static final int LIST_INNER_TYPE_POS = 0;
+
+  public ListMapper(DatatypeMappingHandler resolveInstance) {
+    super(resolveInstance);
+  }
+
+  public boolean applies(TypeName type) {
+    if (type instanceof ParameterizedTypeName) {
+      return ((ParameterizedTypeName) type).rawType.simpleName().equals(List.class.getSimpleName());
+    }
+    return TypeName.get(List.class).equals(type);
+  }
 
   @Override
-  public CodeBlock getReturnStatement(Object resultToReturn) {
-    // @TODO split result into list
+  public boolean appliesToJSON(Object typeDefForResolvingInnerClass) {
+    return parseToJsonArray(typeDefForResolvingInnerClass) != null;
+  }
+
+  public CodeBlock encodeValue(TypeName type, String variableName) {
+    String uniqueVariableName = this.getUniqueVariableName(variableName);
+    TypeName innerType = this.getListInnerType(type);
+
     return CodeBlock.builder()
-        .addStatement("return $T.of($L)", Arrays.class, resultToReturn)
+        .add("$L.stream().map($L->", variableName, uniqueVariableName)
+        .add("$L", this.resolveInstance.encodeParameter(innerType, uniqueVariableName))
+        .add(").collect($T.toList()).toString()", Collectors.class)
+        .build();
+  }
+
+  public CodeBlock mapToReturnValue(TypeName type, String variableName) {
+    String uniqueVariableName = this.getUniqueVariableName(variableName);
+    TypeName innerType = this.getListInnerType(type);
+
+    return CodeBlock.builder()
+        .add(
+            "(($T)$L).stream().map($L->",
+            LIST_WITH_WILDCARD_TYPDEF,
+            variableName,
+            uniqueVariableName)
+        .add("$L", this.resolveInstance.mapToReturnValue(innerType, uniqueVariableName))
+        .add(").collect($T.toList())", Collectors.class)
         .build();
   }
 
   @Override
-  public Type getJavaType() {
-    return List.class;
+  public TypeName getTypeNameFromJSON(Object type) {
+    return ParameterizedTypeName.get(
+        ClassName.get(List.class),
+        this.resolveInstance.getTypeNameFromJSON(
+            parseToJsonArray(type).getValue(LIST_INNER_TYPE_POS)));
   }
 
-  @Override
-  public boolean applies(Object type) {
-    try {
-      return JsonObject.mapFrom(type).containsKey("list");
-    } catch (Exception e) {
-      // ignore because if it is not a json we do not need to go on
-      // further
+  private JsonArray parseToJsonArray(Object type) {
+    if (type != null) {
+      if (type instanceof JsonObject) {
+        Object list = JsonObject.mapFrom(type).getValue(LIST_JSON_IDENTIFIER);
+        if (list instanceof JsonArray) {
+          return (JsonArray) list;
+        }
+      }
     }
-    return false;
+    return null;
+  }
+
+  private TypeName getListInnerType(TypeName type) {
+    if (type instanceof ParameterizedTypeName) {
+      return ((ParameterizedTypeName) type).typeArguments.get(LIST_INNER_TYPE_POS);
+    }
+    throw new RuntimeException(
+        getUnforseenMappingMessage(
+            "given argument for resolving list type seems to be not parametrized",
+            ListMapper.class.getName(),
+            "getListInnerType",
+            type.toString()));
   }
 }